train/requirements.txt
appdirs==1.4.4
audioread==2.1.9
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
decorator==5.0.9
fastprogress==1.0.0
google-api-core==2.0.0
google-auth==2.0.1
google-cloud-texttospeech==2.5.2
googleapis-common-protos==1.53.0
grpcio==1.39.0
idna==3.2
joblib==1.0.1
librosa==0.8.1
llvmlite==0.37.0
numba==0.54.0
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pooch==1.5.1
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
SoundFile==0.10.3.post1
TextGrid==1.5
threadpoolctl==2.2.0
tqdm==4.62.2
urllib3==1.26.6
